CHG: refine file name pythonically; move s4 to layers; update makefile based on the file modification
