+++ sssd/core/layers/s4/s4_layer.py
@@ -6,7 +6,7 @@ from einops import rearrange
 
 from sssd.core.layers.activation import Activation
 from sssd.core.layers.linear import LinearActivation
-from sssd.core.s4.hippo.hippo import HippoSSKernel
+from sssd.core.layers.s4.hippo.hippo import HippoSSKernel
 from sssd.utils.logger import setup_logger
 
 contract = oe.contract
